:recycle: [Refactor]: 방송 제목 컴포넌트 분리
