+++ apps/client/src/pages/Broadcast.tsx
@@ -1,3 +1,4 @@
+import BroadcastTitle from '@/components/BroadcastTitle';
 import { useMediaControls } from '@/hooks/useMediaControls';
 import { useMediaStream } from '@/hooks/useMediaStream';
 import { useProducer } from '@/hooks/useProducer';
@@ -12,13 +13,8 @@ import VideoOffIcon from '@components/icons/VideoOffIcon';
 import VideoOnIcon from '@components/icons/VideoOnIcon';
 import { Button } from '@components/ui/button';
 import { useCallback, useEffect, useState } from 'react';
-import { useForm, SubmitHandler } from 'react-hook-form';
 
 const socketUrl = import.meta.env.VITE_MEDIASERVER_URL;
-
-interface Inputs {
-  title: string;
-}
 
 function Broadcast() {
   const { mediaStream, mediaStreamError, isMediastreamReady, videoRef } = useMediaStream();
@@ -34,13 +30,7 @@ function Broadcast() {
     device,
     roomId,
   });
-  const [isEditing, setIsEditing] = useState(false);
-  const [title, setTitle] = useState<string>('초기 방송 제목');
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<Inputs>();
+  const [title, setTitle] = useState<string>('방송 제목');
 
   const stopBroadcast = useCallback(
     (e?: BeforeUnloadEvent) => {
@@ -74,13 +64,8 @@ function Broadcast() {
     window.close();
   };
 
-  const handleEditTitle = () => {
-    setIsEditing(true);
-  };
-
-  const onSubmit: SubmitHandler<Inputs> = data => {
-    setTitle(data.title);
-    setIsEditing(false);
+  const handleBroadcastTitle = (newTitle: string) => {
+    setTitle(newTitle);
   };
 
   return (
@@ -98,36 +83,7 @@ function Broadcast() {
             <audio />
           </div>
           <div className="w-full">
-            {isEditing ? (
-              <div className="flex flex-row justify-between p-4 w-full h-20">
-                <form onSubmit={handleSubmit(onSubmit)} className="flex w-full">
-                  <div className="flex-1 mr-2 relative">
-                    <input
-                      defaultValue={title}
-                      {...register('title', { required: true, maxLength: 50 })}
-                      className="w-full h-10 bg-transparent border border-default rounded-md focus:border-bold px-3"
-                    />
-                    {(errors.title?.type === 'required' || errors.title?.type === 'maxLength') && (
-                      <p role="alert" className="absolute top-11 left-0 text-text-danger text-display-medium12">
-                        {errors.title?.type === 'required'
-                          ? '방송 제목을 입력해주세요'
-                          : '방송 제목은 50자 이하로 입력해주세요'}
-                      </p>
-                    )}
-                  </div>
-                  <Button type="submit" className="h-10 shrink-0">
-                    저장
-                  </Button>
-                </form>
-              </div>
-            ) : (
-              <div className="flex flex-row justify-between p-4 h-20">
-                <div className="text-text-default text-display-bold24">{title}</div>
-                <Button className="bg-transparent border border-border-default" onClick={handleEditTitle}>
-                  수정
-                </Button>
-              </div>
-            )}
+            <BroadcastTitle currentTitle={title} onTitleChange={handleBroadcastTitle} />
             <div className="flex justify-between items-center m-4">
               <Button onClick={handleCheckout} className="bg-surface-brand-default hover:hover:bg-surface-brand-alt">
                 체크아웃
